refactor: usecase, service 패키지 세분화  - command 패키지 하위로 이동
