+++ src/test/java/com/somemore/recruitboard/service/command/CreateRecruitBoardServiceTest.java
@@ -1,4 +1,4 @@
-package com.somemore.recruitboard.service;
+package com.somemore.recruitboard.service.command;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -9,6 +9,7 @@ import com.somemore.recruitboard.domain.RecruitBoard;
 import com.somemore.recruitboard.domain.VolunteerType;
 import com.somemore.recruitboard.dto.request.RecruitBoardCreateRequestDto;
 import com.somemore.recruitboard.repository.RecruitBoardRepository;
+import com.somemore.recruitboard.service.command.CreateRecruitBoardService;
 import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.time.LocalTime;
